better version in pyproject and change CLI format
Before, uv added package considering too much details in the packages version. Solved this by removing the last part of these versions.

In the CLI, we have output format names "TERM" and "CSV". Term does not seem clear if we want to output as text. Thus, I renamed it as "TEXT".