feat(github_api): stop on ratelimit without API keys
Before, rabbit would retry 3 times if we had a rate limit without an API key since we don't have any information on the reset time in the payload.

Now, we stop the program directly after a rate limit without api key and invite the user to set an API key.